extract metrics reporting
